Module google_factchecktools1_alpha1::api

Structs

ClaimMethods
A builder providing access to all methods supported on claim resources. It is not used directly, but through the FactCheckTools hub.

ClaimSearchCall
Search through fact-checked claims.

FactCheckTools
Central instance to access all FactCheckTools related resource activities

GoogleFactcheckingFactchecktoolsV1alpha1Claim
Information about the claim.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimAuthor
Information about the claim author.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimRating
Information about the claim rating.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimReview
Information about a claim review.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimReviewAuthor
Information about the claim review author.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimReviewMarkup
Fields for an individual ClaimReview element. Except for sub-messages that group fields together, each of these fields correspond those in https://schema.org/ClaimReview. We list the precise mapping for each field.

GoogleFactcheckingFactchecktoolsV1alpha1ClaimReviewMarkupPage
Holds one or more instances of ClaimReview markup for a webpage.

GoogleFactcheckingFactchecktoolsV1alpha1FactCheckedClaimSearchResponse
Response from searching fact-checked claims.

GoogleFactcheckingFactchecktoolsV1alpha1ListClaimReviewMarkupPagesResponse
Response from listing ClaimReview markup.

GoogleFactcheckingFactchecktoolsV1alpha1Publisher
Information about the publisher.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

PageCreateCall
Create ClaimReview markup on a page.

PageDeleteCall
Delete all ClaimReview markup on a page.

PageGetCall
Get all ClaimReview markup on a page.

PageListCall
List the ClaimReview markup pages for a specific URL or for an organization.

PageMethods
A builder providing access to all methods supported on page resources. It is not used directly, but through the FactCheckTools hub.

PageUpdateCall
Update for all ClaimReview markup on a page Note that this is a full update. To retain the existing ClaimReview markup on a page, first perform a Get operation, then modify the returned markup, and finally call Update with the entire ClaimReview markup as the body.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.